requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions>=1.12.0
fastapi
azure-core==1.26.4
haystack-ai
trafilatura